This creates an average subcortical mask for the given inputs subjects. For each subject, we first create a binary mask from the aseg by removing the following structures:
Background 0
Cerebral Cortex 3 42
Cerebral White Matter 2 41
CC 251 252 253 254 255
Cerebellar White Matter 7 46
Ventricles 4 14 15 43 44 72 75 76 24
Choroid Plexus 31 63

This mask is then converted to "talairach" space (meaning, conformed
MNI305 like fsaverage) using the talairach.xfm.  All the subjects are
averaged together, so each voxel contains a number between 0 and 1,
with 1 indicating that it is highly likely that there is a subcortical
gray matter structure at that location.

The purpose of this is to create a mask for doing volume-based
fMRI analysis of subcortical structure (assuming that the
cortical analysis is going to be done on the surface).